+++ src/types/assets.d.ts
@@ -7,3 +7,8 @@ declare module "*.scm" {
   const path: string
   export default path
 }
+
+declare module "*.md" {
+  const text: string
+  export default text
+}